+++ src/frontend/komponenter/Fagsak/Behandlingsresultat/Eøs/UtbetaltAnnetLand/useUtenlandskPeriodeBeløpSkjema.ts
@@ -6,22 +6,27 @@ import { feil, ok, useFelt, useSkjema } from '@navikt/familie-skjema';
 import type { Ressurs } from '@navikt/familie-typer';
 import { byggTomRessurs, RessursStatus } from '@navikt/familie-typer';
 
-import type { IBehandling } from '../../typer/behandling';
-import type { OptionType } from '../../typer/common';
+import { useBehandling } from '../../../../../context/behandlingContext/BehandlingContext';
+import type { IBehandling } from '../../../../../typer/behandling';
+import type { OptionType } from '../../../../../typer/common';
 import type {
     EøsPeriodeStatus,
     IRestUtenlandskPeriodeBeløp,
     IUtenlandskPeriodeBeløp,
     UtenlandskPeriodeBeløpIntervall,
-} from '../../typer/eøsPerioder';
-import type { IIsoMånedPeriode } from '../../utils/dato';
-import { nyIsoMånedPeriode } from '../../utils/dato';
+} from '../../../../../typer/eøsPerioder';
+import type { IIsoMånedPeriode } from '../../../../../utils/dato';
+import { nyIsoMånedPeriode } from '../../../../../utils/dato';
 import {
     konverterDesimalverdiTilSkjemaVisning,
     konverterSkjemaverdiTilDesimal,
-} from '../../utils/eøs';
-import { erBarnGyldig, erEøsPeriodeGyldig, isEmpty, isNumeric } from '../../utils/eøsValidators';
-import { useBehandling } from '../behandlingContext/BehandlingContext';
+} from '../../../../../utils/eøs';
+import {
+    erBarnGyldig,
+    erEøsPeriodeGyldig,
+    isEmpty,
+    isNumeric,
+} from '../../../../../utils/eøsValidators';
 
 const erBeløpGyldig = (felt: FeltState<string | undefined>): FeltState<string | undefined> => {
     if (!felt.verdi || isEmpty(felt.verdi) || typeof felt.verdi != 'string') {
